+++ core/database/src/main/kotlin/com/unifest/android/core/database/LikedFestivalDao.kt
@@ -14,7 +14,7 @@ interface LikedFestivalDao {
     suspend fun insertLikedFestival(userInfo: LikedFestivalEntity)
 
     @Query("SELECT EXISTS(SELECT 1 FROM liked_festival WHERE festival_id = :festivalId)")
-    suspend fun isFestivalExists(festivalId: Int): Boolean
+    suspend fun isFestivalExists(festivalId: Long): Boolean
 
     @Delete
     suspend fun deleteLikedFestival(userInfo: LikedFestivalEntity)
