+++ core/model/src/main/kotlin/com/unifest/android/core/model/FestivalModel.kt
@@ -4,8 +4,8 @@ import androidx.compose.runtime.Stable
 
 @Stable
 data class FestivalModel(
-    val festivalId: Int,
-    val schoolId: Int,
+    val festivalId: Long,
+    val schoolId: Long,
     val thumbnail: String,
     val schoolName: String,
     val festivalName: String,
